Prepare the curve theme for MediaWIki 1.39 (Tested on 1.38)
